+++ frontend/src/pages/Plugins/index.ts
@@ -2,5 +2,4 @@ export { PluginsHome } from "./Home";
 export { PluginsTemplatesPage } from "./Templates";
 export { PluginsAliasesPage } from "./Aliases";
 export { PluginsSchedulerPage } from "./Scheduler";
-export { PluginsLegacyExtensionsPage } from "./LegacyExtensions";
 export { FEATURE_CONFIG_PAGE_KEYS, featureConfigPath } from "./_shared/featureConfig";